rename reducer files to slices
