+++ src/redux/slices/sectorSlice.ts
